requirements.txt
appdirs==1.4.3
attrs==19.3.0
awscrt==0.5.13
awsiotsdk==1.0.6
black==19.10b0
click==7.1.2
colorzero==1.1
entrypoints==0.3
flake8==3.7.9
gpiozero==1.5.1
mccabe==0.6.1
pathspec==0.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
regex==2020.5.7
toml==0.10.0
typed-ast==1.4.1
